refactor: Extract util for generating hapiPath
