+++ src/utils/index.js
@@ -13,6 +13,7 @@ export { default as parseQueryStringParameters } from './parseQueryStringParamet
 export { default as satisfiesVersionRange } from './satisfiesVersionRange.js'
 export { default as splitHandlerPathAndName } from './splitHandlerPathAndName.js'
 export { default as checkDockerDaemon } from './checkDockerDaemon.js'
+export { default as generateHapiPath } from './generateHapiPath.js'
 // export { default as baseImage } from './baseImage.js'
 
 // Detect the toString encoding from the request headers content-type
